Introduce new build system
This build system is very different to the old build system, and is much
more inuitive.

It leverages a hand made configure script to generate the Makefile. The
listener is no longer auto-detected, unless specifically specified, and
no longer activated by default. That is not the place for a build system.
Along with the focus of moving to distribution repositiories, this 'feature'
is clearly unwanted by distributions.

Multiple listeners are now also supported, primarily for distributions.

Lastly, parallelism now works as well.