fix all call to die() with NULL as an argument
This fixes all the calls to die so they pass a proper string and not
NULL.  die blindly pass its arguments to log_fatal, which in turns
relays on vprintf or such.  At least on OpenBSD, giving NULL instead
of a string for %s yields a log entry in syslog:

Jan 21 09:35:22 venera gophernicus: vfprintf %s NULL in "Error: %s %s\^M "

As a side effect, now some errors from pledge and unveil are slightly
more clear.