Update menu.c: FLAC, MIDI, matroska/webm, and some MPEG format support
Please note that most MPEG file formats are containers (IIRC almost every one that *can* contain video) and there are some irritating issues assigning the correct gopher type (and if we ever adopt gopher+, also assigning the correct MIME type) to these files. In the end it's up to the client to deal with this.

I'm going to err on the side of presenting these to the client as video.